feat: redo component structure for roles (wip)
